+++ R/pal-dichromat.r
@@ -12,6 +12,12 @@
 #' cols <- dichromat_pal("DarkRedtoBlue.12")(12)
 #' show_col(gradient_n_pal(cols)(seq(0, 1, length.out = 30)))
 dichromat_pal <- function(name) {
+  if (!requireNamespace("dichromat", quietly = TRUE)) {
+    stop("Package dichromat must be installed for this function to work. Please install it.",
+      call. = FALSE
+    )
+  }
+
   if (!any(name == names(dichromat::colorschemes))) {
     stop("Palette name must be one of ",
       paste0(names(dichromat::colorschemes), collapse = ", "),