+++ tests/testthat/test-formatter.r
@@ -1,5 +1,7 @@
 context("Formatters")
 
+# Time formatter --------------------------------------------------------
+
 test_that("time_format formats hms objects", {
   a_time <- ISOdatetime(2012, 1, 1, 11, 30, 0, tz = "UTC")
 
@@ -8,11 +10,46 @@ test_that("time_format formats hms objects", {
   expect_equal(time_format(format = "%H")(hms::as.hms(a_time, tz = "UTC")), "11")
 })
 
+# Number formatter --------------------------------------------------------
+
+test_that("number format works correctly", {
+  expect_equal(number(123.45, accuracy = 1), "123")
+  expect_equal(number(123.45, accuracy = 10), "120")
+  expect_equal(number(123.45, accuracy = .25), "123.5")
+  expect_equal(
+    number(12345, big.mark = ","),
+    "12,345"
+  )
+  expect_equal(
+    number(12.3, decimal.mark = ",", accuracy = .1),
+    "12,3"
+  )
+  expect_equal(
+    number(1.234, scale = 100),
+    "123"
+  )
+  expect_equal(
+    number(123, prefix = "pre", suffix = "post"),
+    "pre123post"
+  )
+  expect_equal(number(c(1, 23)), c("1", "23"))
+  expect_equal(number(c(1, 23), trim = FALSE), c(" 1", "23"))
+})
+
+test_that("number_format works with Inf", {
+  cust <- number_format(suffix = "suff", accuracy = NULL)
+  expect_equal(cust(c(Inf, -Inf)), c("Inf", "-Inf"))
+})
+
+# Comma formatter --------------------------------------------------------
+
 test_that("comma format always adds commas", {
   expect_equal(comma(1e3), "1,000")
   expect_equal(comma(1e6), "1,000,000")
   expect_equal(comma(1e9), "1,000,000,000")
 })
+
+# Scientific formatter --------------------------------------------------------
 
 test_that("scientific format shows specific sig figs", {
   expect_equal(scientific(123456, digits = 1), "1e+05")
@@ -23,6 +60,22 @@ test_that("scientific format shows specific sig figs", {
   expect_equal(scientific(0.123456, digits = 2), "1.2e-01")
   expect_equal(scientific(0.123456, digits = 3), "1.23e-01")
 })
+
+test_that("prefix and suffix works with scientific format", {
+  expect_equal(scientific(123456, digits = 2, prefix = "V="), "V=1.2e+05")
+  expect_equal(scientific(123456, digits = 2, suffix = " km"), "1.2e+05 km")
+})
+
+test_that("scale works with scientific format", {
+  expect_equal(scientific(123456, digits = 2, scale = 1000), "1.2e+08")
+})
+
+test_that("decimal.mark works with scientific format", {
+  expect_equal(scientific(123456, digits = 2, decimal.mark = ","), "1,2e+05")
+})
+
+
+# Wrap formatter --------------------------------------------------------
 
 test_that("wrap correctly wraps long lines", {
   expect_equal(
@@ -41,6 +94,8 @@ test_that("wrap correctly wraps long lines", {
   expect_equal(wrap_format(15)("short line"), "short line")
 })
 
+# Ordinal formatter --------------------------------------------------------
+
 test_that("ordinal format", {
   expect_equal(ordinal(1), "1st")
   expect_equal(ordinal(2), "2nd")
@@ -55,33 +110,7 @@ test_that("ordinal format maintains order", {
   expect_equal(ordinal(c(21, 1, 11)), c("21st", "1st", "11th"))
 })
 
-test_that("formatters don't add extra spaces", {
-  has_space <- function(x) any(grepl("\\s", x))
-  x <- 10^c(-1, 0, 1, 3, 6, 9)
-
-  expect_false(has_space(comma(x)))
-  expect_false(has_space(dollar(x)))
-  expect_false(has_space(percent(x)))
-  expect_false(has_space(percent(x)))
-  expect_false(has_space(scientific(x)))
-})
-
-test_that("formats work with 0 length input", {
-  x <- numeric()
-  expected <- character()
-  expect_identical(comma(x), expected)
-  expect_identical(dollar(x), expected)
-  expect_identical(percent(x), expected)
-  expect_identical(scientific(x), expected)
-  expect_identical(comma_format()(x), expected)
-  expect_identical(date_format()(as.Date(character(0))), expected)
-  expect_identical(dollar_format()(x), expected)
-  expect_identical(math_format()(x), list())
-  expect_identical(parse_format()(x), list())
-  expect_identical(percent_format()(x), expected)
-  expect_identical(scientific_format()(x), expected)
-  expect_identical(trans_format(identity)(x), expected)
-})
+# Unit formatter --------------------------------------------------------
 
 test_that("unit format", {
   expect_equal(
@@ -89,20 +118,19 @@ test_that("unit format", {
     c("1 km", "2 km")
   )
   expect_equal(
-    unit_format(unit = "ha", scale = 1e-4)(c(1e3, 2e3)),
+    unit_format(unit = "ha", scale = 1e-4, accuracy = .1)(c(1e3, 2e3)),
     c("0.1 ha", "0.2 ha")
   )
   expect_equal(
-    unit_format()(c(1e3, 2e3)),
-    c("1,000 m", "2,000 m")
+    unit_format()(c(1e2, 2e2)),
+    c("100 m", "200 m")
   )
 })
-
 
 # Percent formatter -------------------------------------------------------
 
 test_that("negative percents work", {
-  expect_equal(percent(-0.6), "-60%")
+  expect_equal(percent(-0.6, accuracy = 1), "-60%")
 })
 
 test_that("Single 0 gives 0%", {
@@ -117,4 +145,35 @@ test_that("negative comes before prefix", {
 
 test_that("missing values preserved", {
   expect_equal(dollar(NA_real_), "$NA")
+})
+
+# Common tests --------------------------------------------------------
+
+test_that("formatters don't add extra spaces", {
+  has_space <- function(x) any(grepl("\\s", x))
+  x <- 10^c(-1, 0, 1, 3, 6, 9)
+
+  expect_false(has_space(number(x, big.mark = ",")))
+  expect_false(has_space(comma(x)))
+  expect_false(has_space(dollar(x)))
+  expect_false(has_space(percent(x, big.mark = ",")))
+  expect_false(has_space(scientific(x)))
+})
+
+test_that("formats work with 0 length input", {
+  x <- numeric()
+  expected <- character()
+  expect_identical(number(x), expected)
+  expect_identical(comma(x), expected)
+  expect_identical(dollar(x), expected)
+  expect_identical(percent(x), expected)
+  expect_identical(scientific(x), expected)
+  expect_identical(comma_format()(x), expected)
+  expect_identical(date_format()(as.Date(character(0))), expected)
+  expect_identical(dollar_format()(x), expected)
+  expect_identical(math_format()(x), list())
+  expect_identical(parse_format()(x), list())
+  expect_identical(percent_format()(x), expected)
+  expect_identical(scientific_format()(x), expected)
+  expect_identical(trans_format(identity)(x), expected)
 })